Add files via upload
sick_beetz_GUI.py is a GUI version of a pygame based music player that I created. sick_beetz_v2.py is a simpler version that works on a terminal. In order for both of them to work, they must be put inside a folder that contains music files. For now the only file types supported are .wav and .mp3